command-line-tools/occurrence/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "occurrence"
dynamic = ["version"]
description = 'Water Bodies Occurrences'

readme = "README.md"
requires-python = ">=3.12,<3.13"
license = "MIT"
keywords = []
authors = [
  { name = "Fabrice Brito", email = "info@terradue.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "zarr",
  "loguru",
  "click",
  "rioxarray",
  "rio-stac",
  "odc-stac",
]

[project.urls]
Documentation = "https://eoap.github.io/zarr-cloud-native-format/"
Issues = "https://github.com/eoap/zarr-cloud-native-format/issues"
Source = "https://github.com/eoap/zarr-cloud-native-format"

[tool.hatch.envs.prod]
path = "/app/envs/prod"

[tool.hatch.version]
path = "occurrence/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["occurrence"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]

[project.scripts]
occurrence = "occurrence.app:occurrence"

[tool.hatch.envs.dev]
description = "Development environment"
dependencies = [
  "ruff",
]

[tool.hatch.envs.dev.scripts]
lint = "ruff format ."
check = "ruff check --fix"
